Addition cleanup (#65)
* Upgrade non-breaking dependencies

* Upgrade ESLint, actually use it

This package had ESLint as a dev-dependency, but, was never
actually using it since there was no `.eslintrc` or equivalent
file. This adds the config and modifies the codebase accordingly.

* Move `react-dom` to `peerDependencies`

We can allow the user to specify their version of `react-dom` which
will hopefully let them ship less code.

* Turn off import/import/no-unresolved

I can't get this rule to work for some reason. Will revist it.

* Rename index.js -> index.jsx

* Add `react` and `react-dom` to `devDependencies`

So we can develop without needing to manually install (which
touches `package-lock.json`, even when not using the `--save` flag
which is beyond frustrating.

* Use .js, not .jsx